fs: large files support, improve SSH sync performance

+++ filediff-node-module/src/main/java/org/sudu/experiments/js/node/JsSftpClient.java
@@ -64,13 +64,13 @@ public interface JsSftpClient extends JSObject {
   void read(
       JSObject handle,
       JsBuffer buffer, int offset,
-      int length, int position, ReadResult cb);
+      int length, double position, ReadResult cb);
 
   // Writes length bytes from buffer starting at offset to the resource
   // associated with handle starting at position.
   void write(JSObject handle,
              JsBuffer buffer, int offset, int length,
-             int position,  JsFunctions.Consumer<JSError> cb);
+             double position, JsFunctions.Consumer<JSError> cb);
 
   // Closes the resource associated with handle given by open() or opendir().
   void close(JSObject handle, JsFunctions.Consumer<JSError> cb);
